// apps/mango-crem-features/list-pages/src/app/components/listpage/shared/models/marker.ts
export interface Marker {
  objectId: number;
  objectTypeId: number;
  objectTypeTypeId: number;
  name: string;
  address1: string;
  address2: string;
  country: string;
  goToUrl: string;
  zoomOnMapUrl: string;
  latitude: number;
  longitude: number;
  redirectorUrl: string;
  htmlString: string;
}